[refactor] Move the share file listing out of the worker entrypoint
listOverlayShareFiles and its row builder move to a shared module with
their data-layer calls injected, so the listing's read pattern becomes
assertable — a closure in the entrypoint has no seam a test can reach.
The entrypoint's duplicate pathFromMount goes with it; both call sites
now use the canonical guard in transfer/path-guard.js. No behaviour
change, and no test changes.
